Move value declaration psi class to declaration package
