+++ src/main/java/org/purescript/file/PSFile.kt
@@ -4,7 +4,7 @@ import com.intellij.extapi.psi.PsiFileBase
 import com.intellij.openapi.fileTypes.FileType
 import com.intellij.psi.FileViewProvider
 import org.purescript.psi.PSModule
-import org.purescript.psi.PSValueDeclaration
+import org.purescript.psi.declaration.PSValueDeclaration
 
 class PSFile(viewProvider: FileViewProvider) :
     PsiFileBase(viewProvider, org.purescript.PSLanguage.INSTANCE) {
