+++ src/main/java/org/purescript/ide/refactoring/PSRefactoringSupportProvider.kt
@@ -3,7 +3,7 @@ package org.purescript.ide.refactoring
 import com.intellij.lang.refactoring.RefactoringSupportProvider
 import com.intellij.psi.PsiElement
 import org.purescript.psi.PSModule
-import org.purescript.psi.PSValueDeclaration
+import org.purescript.psi.declaration.PSValueDeclaration
 
 class PSRefactoringSupportProvider : RefactoringSupportProvider() {
     override fun isMemberInplaceRenameAvailable(element: PsiElement, context: PsiElement?): Boolean {
