+++ src/main/java/org/purescript/psi/PSLetImpl.kt
@@ -1,6 +1,7 @@
 package org.purescript.psi
 
 import com.intellij.lang.ASTNode
+import org.purescript.psi.declaration.PSValueDeclaration
 
 class PSLetImpl(node: ASTNode) : PSPsiElement(node) {
     val valueDeclarations: Array<PSValueDeclaration> get() =
