+++ src/main/java/org/purescript/psi/dostmt/PSDoNotationLetImpl.kt
@@ -2,7 +2,7 @@ package org.purescript.psi.dostmt
 
 import com.intellij.lang.ASTNode
 import org.purescript.psi.PSPsiElement
-import org.purescript.psi.PSValueDeclaration
+import org.purescript.psi.declaration.PSValueDeclaration
 
 class PSDoNotationLetImpl(node: ASTNode) : PSPsiElement(node) {
     val valueDeclarations get() =
