+++ src/main/java/org/purescript/psi/expression/PSExpressionWhere.kt
@@ -2,7 +2,7 @@ package org.purescript.psi.expression
 
 import com.intellij.lang.ASTNode
 import org.purescript.psi.PSPsiElement
-import org.purescript.psi.PSValueDeclaration
+import org.purescript.psi.declaration.PSValueDeclaration
 
 /**
  * A where claus un a expression, e.g.
